Move MoreRoute.kt to common module
